Thu, Sep 29, 2022, 9:16 AM +00:00

+++ cs50x/finance/app.py
@@ -69,7 +69,7 @@ def buy():
         elif not request.form.get("shares"):
             return apology("must provide a valid quantity of shares to buy", 400)
 
-        elif int(request.form.get("shares"))<=0:
+        elif float(request.form.get("shares"))<=0:
             return apology("provide a valid share input",400)
 
         elif (float(request.form.get("shares")) % 1) != 0:
@@ -232,7 +232,7 @@ def sell():
             return apology("must provide a valid stock symbol", 400)
 
         # Ensure password was submitted
-        elif not request.form.get("shares") or request.form.get("shares")<=0:
+        elif not request.form.get("shares") or float(request.form.get("shares"))<=0:
             return apology("must provide a valid quantity of shares to buy", 400)
 
         elif (float(request.form.get("shares")) % 1) != 0:
